Make some code tweaks
